refactor: BonusNumber 도메인의 number 타입을 long에서 int로 변경
- BonusNumber 클래스의 number 필드 타입을 int로 변경
- 생성자와 from 메서드의 파라미터 타입을 int로 변경

+++ src/main/java/lotto/controller/LottoController.java
@@ -57,7 +57,7 @@ public class LottoController {
     private BonusNumber inputBonus() {
         try {
             long winningBonusNumberInput = InputView.inputNumber(InputMessage.BONUS_NUMBER);
-            return BonusNumber.from(winningBonusNumberInput);
+            return BonusNumber.from((int)winningBonusNumberInput);
         } catch (IllegalArgumentException e) {
             OutputView.printException(e);
             return inputBonus();
